chore: file restructure
