+++ cli/src/commands/root.go
@@ -1,7 +1,8 @@
-package cmd
+package commands
 
 import (
 	"fmt"
+	envCommand "horahora/cli/src/commands/env"
 	"os"
 
 	"github.com/spf13/cobra"
@@ -46,7 +47,7 @@ func init() {
 	// Cobra also supports local flags, which will only run
 	// when this action is called directly.
 	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-	rootCmd.AddCommand(envCmd)
+	rootCmd.AddCommand(envCommand.ENVCommand)
 }
 
 func initConfig() {
